restructured for better usecase
