+++ src/pages/index.js
@@ -1,13 +1,14 @@
-import Banner from "@/component/Banner"
-import RootLayout from "@/layouts/RootLayout"
+import Banner from "@/components/Banner/Banner";
+import RootLayout from "@/layouts/RootLayout";
 
 export default function Home() {
-    
-    return <div>
-        <Banner></Banner>
+  return (
+    <div>
+      <Banner></Banner>
     </div>
+  );
 }
 
 Home.getLayout = (page) => {
-    return <RootLayout>{page}</RootLayout>
-}
+  return <RootLayout>{page}</RootLayout>;
+};
